+++ specifyweb/frontend/js_src/lib/components/FormMeta/index.tsx
@@ -18,7 +18,6 @@ import { PrintOnSave } from '../FormFields/Checkbox';
 import type { ViewDescription } from '../FormParse';
 import { SubViewContext } from '../Forms/SubView';
 import { isTreeResource } from '../InitialContext/treeRanks';
-import { interactionTables } from '../Interactions/InteractionsDialog';
 import { Dialog } from '../Molecules/Dialog';
 import {
   ProtectedAction,
@@ -34,6 +33,7 @@ import { QueryTreeUsages } from './QueryTreeUsages';
 import { ReadOnlyMode } from './ReadOnlyMode';
 import { ShareRecord } from './ShareRecord';
 import { SubViewMeta } from './SubViewMeta';
+import { interactionTables } from '../Interactions/config';
 
 /**
  * Form preferences host context aware user preferences and other meta-actions.
