+++ specifyweb/frontend/js_src/lib/components/Interactions/InteractionsDialog.tsx
@@ -40,45 +40,6 @@ import { formatUrl } from '../Router/queryString';
 import { OverlayContext } from '../Router/Router';
 import { InteractionDialog } from './InteractionDialog';
 
-export const interactionTables: ReadonlySet<keyof Tables> = new Set<
-  keyof Tables
->([
-  'Accession',
-  'AccessionAgent',
-  'AccessionAttachment',
-  'AccessionAuthorization',
-  'AccessionCitation',
-  'Appraisal',
-  'Borrow',
-  'BorrowAgent',
-  'BorrowAttachment',
-  'BorrowMaterial',
-  'BorrowReturnMaterial',
-  'Deaccession',
-  'DeaccessionAgent',
-  'DeaccessionAttachment',
-  'Disposal',
-  'DisposalAgent',
-  'DisposalAttachment',
-  'DisposalPreparation',
-  'ExchangeIn',
-  'ExchangeInPrep',
-  'ExchangeOut',
-  'ExchangeOutPrep',
-  'Gift',
-  'GiftAgent',
-  'GiftAttachment',
-  'GiftPreparation',
-  'InfoRequest',
-  'Loan',
-  'LoanAgent',
-  'LoanAttachment',
-  'LoanPreparation',
-  'LoanReturnPreparation',
-  'Permit',
-  'PermitAttachment',
-]);
-
 const supportedActions = [
   'NEW_GIFT',
   'NEW_LOAN',
