+++ specifyweb/frontend/js_src/lib/components/Interactions/config.ts
@@ -1,0 +1,40 @@
+import { Tables } from '../DataModel/types';
+
+export const interactionTables: ReadonlySet<keyof Tables> = new Set<
+  keyof Tables
+>([
+  'Accession',
+  'AccessionAgent',
+  'AccessionAttachment',
+  'AccessionAuthorization',
+  'AccessionCitation',
+  'Appraisal',
+  'Borrow',
+  'BorrowAgent',
+  'BorrowAttachment',
+  'BorrowMaterial',
+  'BorrowReturnMaterial',
+  'Deaccession',
+  'DeaccessionAgent',
+  'DeaccessionAttachment',
+  'Disposal',
+  'DisposalAgent',
+  'DisposalAttachment',
+  'DisposalPreparation',
+  'ExchangeIn',
+  'ExchangeInPrep',
+  'ExchangeOut',
+  'ExchangeOutPrep',
+  'Gift',
+  'GiftAgent',
+  'GiftAttachment',
+  'GiftPreparation',
+  'InfoRequest',
+  'Loan',
+  'LoanAgent',
+  'LoanAttachment',
+  'LoanPreparation',
+  'LoanReturnPreparation',
+  'Permit',
+  'PermitAttachment',
+]);
