+++ specifyweb/frontend/js_src/lib/components/Preferences/Editor.tsx
@@ -1,0 +1,40 @@
+import React from 'react';
+
+import { PreferencesContent } from '../Preferences';
+import { BasePreferences } from '../Preferences/BasePreferences';
+import { userPreferenceDefinitions } from '../Preferences/UserDefinitions';
+import { userPreferences } from '../Preferences/userPreferences';
+import { AppResourceTab } from '../AppResources/TabDefinitions';
+import { useLiveState } from '../../hooks/useLiveState';
+
+export const UserPreferencesEditor: AppResourceTab = function ({
+  isReadOnly,
+  data,
+  onChange: handleChange,
+}): JSX.Element {
+  const [preferencesContext] = useLiveState<typeof userPreferences>(
+    React.useCallback(() => {
+      const userPreferences = new BasePreferences({
+        definitions: userPreferenceDefinitions,
+        values: {
+          resourceName: 'UserPreferences',
+          fetchUrl: '/context/user_resource/',
+        },
+        defaultValues: undefined,
+        developmentGlobal: '_editingUserPreferences',
+        syncChanges: false,
+      });
+      userPreferences.setRaw(
+        JSON.parse(data === null || data.length === 0 ? '{}' : data)
+      );
+      return userPreferences;
+    }, [handleChange])
+  );
+
+  const Context = userPreferences.Context;
+  return (
+    <Context.Provider value={preferencesContext}>
+      <PreferencesContent isReadOnly={isReadOnly} />
+    </Context.Provider>
+  );
+};
